% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods_correlation_distance.R
\name{get_methods_correlation_distance}
\alias{get_methods_correlation_distance}
\title{Show the methods for calculating correlation or distance}
\usage{
get_methods_correlation_distance(names = TRUE)
}
\arguments{
\item{names}{whether to output names (default) or characterisation
as similarity or dissimilarity (used internally to invert dissimilarity measures)}
}
\value{
A character vector of options for the method arguement of calculate_expression_similarity_*();
if names=FALSE, a vector of types (similarity/dissimilarity measure) of the same length
}
\description{
This function outputs the methods available for the calculation of
the correlation or distance. The standard correlation methods use stats::cor and
a wide variety of distance methods are available using the philentropy package.
To be used as input in calculate_expression_similarity_*().
}
\examples{
get_methods_correlation_distance()
}
