% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisyr_counts.R
\name{noisyr_counts}
\alias{noisyr_counts}
\title{Run the noisyR pipeline for the count matrix approach}
\usage{
noisyr_counts(
  expression.matrix,
  n.elements.per.window = NULL,
  optimise.window.length.logical = FALSE,
  similarity.threshold = 0.25,
  method.chosen = "Boxplot-IQR",
  minimise.coefficient.of.variation = FALSE,
  ...
)
}
\arguments{
\item{expression.matrix}{the expression matrix used as input for the similarity calculation;
this argument is required}

\item{n.elements.per.window}{number of elements to have in a window passed to
calculate_expression_similarity_counts(); default 10\% of the number of rows}

\item{optimise.window.length.logical}{whether to call optimise_window_length to try and
optimise the value of n.elements.per.window}

\item{similarity.threshold, method.chosen}{parameters passed on to calculate_noise_threshold_base()}

\item{minimise.coefficient.of.variation}{whether to compute optimal values for similarity.threshold
and method.chosen by minimising the coefficient of variation across samples}

\item{...}{arguments to be passed on to individual pipeline steps; see their documentation
for more details and required arguments}
}
\value{
For the counts approach, the denoised expression matrix. For the transcript approach,
the numeric vector of noise thresholds per sample. For more details, see their respective documentation.
}
\description{
Calls the functions to run each of the three steps of the pipeline
(similarity calculation, noise quantification, noise removal), with the specified parameters.
}
\examples{
noisyr_counts(
    expression.matrix = matrix(1:100, ncol = 5),
    similarity.measure = "correlation_pearson",
    n.elements.per.window = 3)
}
\seealso{
[noisyr()], [noisyr_transcript()]
}
