% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise_window_length.R
\name{optimise_window_length}
\alias{optimise_window_length}
\title{Calcualte the distance matrices using the count matrix}
\usage{
optimise_window_length(
  expression.matrix,
  method = "correlation_pearson",
  winlen.min = NULL,
  winlen.max = NULL,
  winlen.by = NULL,
  nstep.frac = 0.05,
  iternum = 1000,
  save.plot = NULL
)
}
\arguments{
\item{expression.matrix}{expression matrix, can be normalized or not}

\item{method}{one of the correlation or distance metrics to be used,
defaults to pearson correlation; list of all methods in
get_methods_correlation_distance()}

\item{nstep.frac}{an alternative way to specify the step size, as a fraction of
the window length; default is 5%}

\item{n.elements.per.window}{number of elements to have in a window,
default 10\% of the number of rows}

\item{nstep}{step size to slide across, default 1\% of n.elements.per.window}
}
\value{
A single value of the optimal number of elements per window
}
\description{
This function generates an average correlation/distance coefficient
for every sliding window, for each sample in the expression matrix.
That is done by comparing the distribution of genes in each window across samples
}
\examples{

}
