% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_methods_calculate_noise_threshold.R
\name{get_methods_calculate_noise_threshold}
\alias{get_methods_calculate_noise_threshold}
\title{Show the methods for calculating a noise threshold}
\usage{
get_methods_calculate_noise_threshold()
}
\value{
A character vector of options for the method.chosen arguement of
\code{\link{calculate_noise_threshold}}
}
\description{
This function outputs the methods available for the calculation of
the noise threshold. To be used as input in \code{\link{calculate_noise_threshold}}.
}
\examples{
get_methods_calculate_noise_threshold()
}
