% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisyr.R
\name{noisyr}
\alias{noisyr}
\title{Run the noisyR pipeline}
\usage{
noisyr(approach_for_similarity_calculation = c("counts", "transcript"), ...)
}
\arguments{
\item{approach_for_similarity_calculation}{which approach to use for the similarity calculation;
defaults to counts}

\item{...}{arguments to be passed on to noisyr_counts or noisyr_transcript; see their documentation
for more details and required arguments}
}
\value{
For the counts approach, the denoised expression matrix. For the transcript approach,
the numeric vector of noise thresholds per sample. For more details, see their respective documentation.
}
\description{
Calls one of noisyr_counts or noisyr_transcript, with the specified parameters.
}
\examples{
noisyr(approach_for_similarity_calculation = "counts",
       expression.matrix = matrix(1:100, ncol = 5))
}
\seealso{
[noisyr_counts()], [noisyr_transcript()]
}
